`__getParentBatches()` : Also match on context
If only matching batches to job / batch combinations based on the node,
parent batches would ignore context variations, resulting in incorrect
dependency configurations in some cases. One example is a task
downstream of a Wedge node that would only depend on one of the context
variations of the task being wedged.